Fixed warning in ScintillaWin. makefile_bor uses Lex*.

+++ win32/ScintillaWin.cxx
@@ -527,7 +527,7 @@ bool ScintillaWin::ModifyScrollBars(int nMax, int nPage) {
 	sci.fMask = SIF_PAGE | SIF_RANGE;
 	BOOL bz = ::GetScrollInfo(wMain.GetID(), SB_VERT, &sci);
 	if ((sci.nMin != 0) || (sci.nMax != pdoc->LinesTotal()) ||
-	        (sci.nPage != (pdoc->LinesTotal() - MaxScrollPos() + 1)) ||
+	        (sci.nPage != static_cast<unsigned int>(pdoc->LinesTotal() - MaxScrollPos() + 1)) ||
 	        (sci.nPos != 0)) {
 		//Platform::DebugPrintf("Scroll info changed %d %d %d %d %d\n",
 		//	sci.nMin, sci.nMax, sci.nPage, sci.nPos, sci.nTrackPos);
